Tone mapping with EnergyMap
After reading the EMap.mat file, the output is the struct type, so I changed it to matrix. So that the function tonemap can read it as an input.
There are still some problems.
First, in tone mapping, we will get the warning message that the image is too big.
Second, the two output images(form local and global in tonemap.m) are both black.